pyproject.toml
[project]
name = "claudecode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "langchain-anthropic>=1.4.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.0.0",
    "reportlab>=4.0.0",
    "anthropic>=0.28.0",
    "python-dotenv>=1.2.2",
]
